requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
boto3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.2
django-filter==25.1
djangorestframework==3.16.0
django-storages
et_xmlfile==2.0.0
Faker==37.3.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.38.2
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.1.2
protobuf==6.30.2
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.45.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
whitenoise==6.9.0
xgboost==3.0.0